git-debimport isc-dhcp_4.3.1.orig.tar.gz

+++ omapip/alloc.c
@@ -4,6 +4,7 @@
    protocol... */
 
 /*
+ * Copyright (c) 2012,2014 by Internet Systems Consortium, Inc. ("ISC")
  * Copyright (c) 2009-2010 by Internet Systems Consortium, Inc. ("ISC")
  * Copyright (c) 2004-2007 by Internet Systems Consortium, Inc. ("ISC")
  * Copyright (c) 1999-2003 by Internet Software Consortium
@@ -26,12 +27,6 @@
  *   <info@isc.org>
  *   https://www.isc.org/
  *
- * This software has been written for Internet Systems Consortium
- * by Ted Lemon in cooperation with Vixie Enterprises and Nominum, Inc.
- * To learn more about Internet Systems Consortium, see
- * ``https://www.isc.org/''.  To learn more about Vixie Enterprises,
- * see ``http://www.vix.com''.   To learn more about Nominum, Inc., see
- * ``http://www.nominum.com''.
  */
 
 #include "dhcpd.h"
@@ -311,7 +306,7 @@ void dmalloc_dump_outstanding ()
 				inhistory = 1;
 				if (!noted) {
 				    log_info ("  %s(%d): %ld", dp -> file,
-					      dp -> line, dp -> size);
+					      dp -> line, (long) dp -> size);
 				    noted = 1;
 				}
 				print_rc_hist_entry (i);
@@ -324,7 +319,8 @@ void dmalloc_dump_outstanding ()
 			if (!inhistory)
 #endif
 				log_info ("  %s(%d): %ld",
-					  dp -> file, dp -> line, dp -> size);
+					  dp -> file, dp -> line, 
+					  (long) dp -> size);
 		}
 #endif
 	}
